Apparently the order of input attributes matter...?
The input range's (default) value wasn't valid if it was > the default max of 100 (for mireds), so on re-render it wouldn't set the value on occassion.

However, if you changed the value manually after the initial render, then it would render the correct value.

This bug was haunting me for way too long